enforce the calling package is the default dialer for VisualVoicemail
In the docs for setVisualVoicemailSmsFilterSettings, they state that
the callingPackage needs to be the default dialer. However, server side,
there is no enforcement for this. Now, every client is verified to be
the default dialer, system dialer, or carrier visual voicemail app
before changing the settings for visual voicemail.

Bug: 308932906
Test: CTS
Change-Id: I951d7783f5c425c03768efb7aee7a38e299e6536
Merged-In: I951d7783f5c425c03768efb7aee7a38e299e6536