Add additional condition satellite is being enabled for rejecting e911 call
It needs quite long time to complete enable satellite modem for oem satellite service, there is race condition that satellite is requested to be enabled but not enabled yet and emergency call is attempted.
In that case the call should be rejected but made because satellite is not in enabled state yet.
Introduced additional condition isSatelliteBeingEnabled which set when satellite is requested to be enabled then reset when satellite is enabled.

Bug: 355301942
Flag: EXEMPT bugfix
Test: manually e2e test
      atest TelephonyConnectionServiceTest
Change-Id: Ief292d6191471457a4aea11d9a3766034a855dd0