Ensure VVM receivers are exported.
See discussion in b/352459220 for more detail.  The tl;dr is that on our
internal test platforms IMPLICIT_INTENTS_ONLY_MATCH_EXPORTED_COMPONENTS
ends up being false for packages/services/Telephony so we have never seen
this particular issue.  There are new platform policies which require
receivers to be exported or no broadcasts will be sent to them.

Flag: EXEMPT bugfix; not possible to flag an attribute in a manifest.
Test: Run telephony unit tests and cts tests
Bug: 352459220
Change-Id: Ia37dd371b3a997e928e281fa2c550b9408e06843