Fixed crash in SafetySourceReceiver
The crash is caused by a broadcast intent that is being sent to
the com.android.phone/.security.SafetySourceReceiver component.
The broadcast intent is being sent with the FLAG_RECEIVER_FOREGROUND
flag, which means that the receiver will be started in the foreground.
This is causing the crash because the receiver is not expecting to be
started in the foreground.

Fixed by making the receiver exported.

Fix: 361578269
Test: Manual
Test: Basic telephony functionality tests
Test: atest TeleServicesTest
Flag: EXEMPT bug fix
Change-Id: Idc4d868eef7fc46a3ed9ea90bc3feadfafbefd72